:green_heart: :recycle: Move base.go to runners/base.go and add aliases.go to parsertest
