+++ parser/parsertest/aliases.go
@@ -1,0 +1,13 @@
+package parsertest
+
+import (
+	"github.com/akm/tparser/parser/parsertest/runners"
+)
+
+var (
+	NewTestParser         = runners.NewTestParser
+	NewTestProgramParser  = runners.NewTestProgramParser
+	NewTestUnitParser     = runners.NewTestUnitParser
+	NewTestProgramContext = runners.NewTestProgramContext
+	NewTestUnitContext    = runners.NewTestUnitContext
+)
